+++ src/main/java/ru/skypro/homework/controller/AdsController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.*;
 import ru.skypro.homework.service.AdsService;
+import ru.skypro.homework.service.CommentService;
 
 @RestController
 @RequestMapping("/ads")
@@ -16,9 +17,11 @@ import ru.skypro.homework.service.AdsService;
 public class AdsController {
 
     private final AdsService adsService;
+    private final CommentService commentService;
 
-    public AdsController(AdsService adsService) {
+    public AdsController(AdsService adsService, CommentService commentService) {
         this.adsService = adsService;
+        this.commentService = commentService;
     }
 
     @GetMapping
@@ -41,7 +44,7 @@ public class AdsController {
     @GetMapping("/{ad_pk}/comments")
     public ResponseEntity<ResponseWrapperComment> getComments(@PathVariable(name = "ad_pk", required = true) String adPk) {
         log.info("Start AdsController method getComments");
-        ResponseWrapperComment result = adsService.getAllCommentsForAd(adPk);
+        ResponseWrapperComment result = commentService.getAllCommentsForAd(adPk);
         if (result == null) {
             return ResponseEntity.notFound().build();
         }
@@ -52,7 +55,7 @@ public class AdsController {
     public ResponseEntity<CommentDto> addComments(@PathVariable(name = "ad_pk", required = true) String adPk,
                                                   @RequestBody(required = true) CommentDto commentDto) {
         log.info("Start AdsController method addComments");
-        CommentDto result = adsService.addComments(adPk, commentDto);
+        CommentDto result = commentService.addComments(adPk, commentDto);
         if (result == null) {
             return ResponseEntity.notFound().build();
         }
@@ -96,7 +99,7 @@ public class AdsController {
     public ResponseEntity<CommentDto> getComments(@PathVariable(name = "ad_pk", required = true) String adPk,
                                                   @PathVariable(name = "id", required = true) int id) {
         log.info("Start AdsController method getComments");
-        CommentDto result = adsService.getCommentForAdByCommentId(adPk, id);
+        CommentDto result = commentService.getCommentForAdByCommentId(adPk, id);
         if (result == null) {
             return ResponseEntity.notFound().build();
         }
@@ -107,11 +110,11 @@ public class AdsController {
     public ResponseEntity<Void> deleteComments(@PathVariable(name = "ad_pk", required = true) String adPk,
                                                @PathVariable(name = "id", required = true) int id) {
         log.info("Start AdsController method deleteComments");
-        CommentDto commentDelete = adsService.getCommentForAdByCommentId(adPk, id);
+        CommentDto commentDelete = commentService.getCommentForAdByCommentId(adPk, id);
         if (commentDelete == null) {
             return ResponseEntity.notFound().build();
         }
-        adsService.deleteComments(adPk, id);//adPk нет смысла передавать
+        commentService.deleteComments(adPk, id);//adPk нет смысла передавать
         return ResponseEntity.ok().build();
     }
 
@@ -120,7 +123,7 @@ public class AdsController {
                                                      @PathVariable(name = "id", required = true) int id,
                                                      @RequestBody CommentDto commentDto) {
         log.info("Start AdsController method updateComments");
-        CommentDto result = adsService.updateComments(adPk, id, commentDto);
+        CommentDto result = commentService.updateComments(adPk, id, commentDto);
         if (result == null) {
             return ResponseEntity.notFound().build();
         }
