+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/market
 spring.datasource.username=admin
 spring.datasource.password=market
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=validate
 
-#spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
+spring.liquibase.change-log=classpath:liquibase/changelog-master.yml
